Remove unused assets, pages
Over the years, we have gathered many assets which are no longer used.
Let's remove unused assets and pages. With this, the size of
repository (excluding .git/) reduces to 55.6 MiB from the original 198.3
MiB.

We can reasonably expect the repository to shrink down to 20 MiB after
some agressive (and necessary) image optimization.