Load fewer resources (#91)
* Drop counterup, countdown js libraries

Counterup and Countdown libraries are used for the "countdown until big
event" section on the index page. Loading the libraries on other pages
is wasteful.

Signed-off-by: Abhishek Kumar <abhishekkumar8222@gmail.com>

* Drop wow js library

Wow is used to reveal animations as we scroll. While it might have had a
purpose, we no longer use any of effects.

Signed-off-by: Abhishek Kumar <abhishekkumar8222@gmail.com>

* Drop waypoint js library

Waypoints is the easiest way to trigger a function when you scroll to an
element. The waypoint is used for speakers.html, which itself is no
longer used.

Signed-off-by: Abhishek Kumar <abhishekkumar8222@gmail.com>

* Drop owl carousel

Owl carousel is used in index page to generate "past speakers" carousel
and has no use in other pages.

Signed-off-by: Abhishek Kumar <abhishekkumar8222@gmail.com>

* Serve popper through CDN

Popper is one of js required for bootstrap. Instead of serving locally,
let's use CDN distribution.

Signed-off-by: Abhishek Kumar <abhishekkumar8222@gmail.com>

* Drop google map

We no longer display a map on our contact page. Let's remove associated
initialization code and library.

Signed-off-by: Abhishek Kumar <abhishekkumar8222@gmail.com>

* Drop et-line

et-line is an icon library, which is no longer used.

Signed-off-by: Abhishek Kumar <abhishekkumar8222@gmail.com>

* Drop animate.css

Signed-off-by: Abhishek Kumar <abhishekkumar8222@gmail.com>

* Drop font-awesome for ionicons

We serve font-awesome locally while requesting ionicons. Having two
different icon libraries is redundant, and since font-awesome also
requires linearicons, let's drop it.

Signed-off-by: Abhishek Kumar <abhishekkumar8222@gmail.com>

* Drop Monsterrat font weights 500, 900

Signed-off-by: Abhishek Kumar <abhishekkumar8222@gmail.com>

* Drop parallax library

The parallax effect is easy to replicate using css. Let's drop the
external library and add the effect using css.

We crop the assets/img/bg/team.jpg, to make it look better.

Signed-off-by: Abhishek Kumar <abhishekkumar8222@gmail.com>

* Upgrade to jquery 3.5.1

Due to dependency on parallax.js, many pages used jquery-2.4.1 with
index including both jquery-2.4.1 and jquery-3.4.1. As parallax.js is
replaced with css, let's upgrade the jquery version as well.

Signed-off-by: Abhishek Kumar <abhishekkumar8222@gmail.com>

* Merge lazy_load.js and main.js

Lazy load is included in each of the pages. By merging main and lazy
load scripts, we can reduce 1 request for each page.

Signed-off-by: Abhishek Kumar <abhishekkumar8222@gmail.com>

* remove unused js files

Signed-off-by: Abhishek Kumar <abhishekkumar8222@gmail.com>