Optimize images for format and dimensions
We have storing images without concern to the image dimensions and
format. As we set upper limits on image dimensions by CSS options and
also simply due to practical size constraints. For example, team images
are four in a row, so they cannot be larger than screen width/4 which
turns out to be 300 pixels for most ordinary screens.

With this aggressive culling, we drop the size of image assets from 48
Mb to 15.5 Mb (Fewer by ~67.7%).

There's still further scope for optimization, as we duplicate images
between past-editions and past-speakers. But that would require a
substantial rewrite of javascript generator logic, and we might
migrate to SSG which would such an endeavor redundant.

Signed-off-by: Abhishek Kumar <abhishekkumar8222@gmail.com>